Removed Star Background due to lagging
The website would behave normal with no lagging, even when searching for names. However, as soon as you would open the Dev Tools, the stars would start to lag, so far that even the results would load much longer. Therefore I cut the line and removed every aspect of it, since I could find a workaround.