Backend Pagination + Bug Fixes
I did a whole work over for the pagination process. Before all the names that fit the filter where retrieved from the server, and only a specific amount would get displayed. While this reduced lag in showing the names, it technically took longer than needed to retrieve the names. Now only a set amount of Names are requested from the server, which also only requests a set amount from the DB with the offset and limit functions.

I could also remove the arrows from the pagination index (atleast on Chrome). And limited it minimum value by setting its minimum to 1 and also clamp the value to the lowest / highest value based on the num of pages by using min and max functions in the EventListener.

Removed old functions from Main.mjs.